requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
feedparser==6.0.11
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
numpy==2.0.1
pandas==2.2.2
psycopg2==2.9.9
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
WTForms==3.1.2
